pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic-uplink-rpi5"
version = "0.1.0"
description = "AIoT semantic uplink PoC on Raspberry Pi 5 (edge prediction + policy LinUCB + MQTT + tc + AoI/MAE/Rate)"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name="Project Owner", email="owner@example.com"}]

dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "paho-mqtt>=1.6",
    "msgspec>=0.18",
    "sounddevice>=0.4",
    "aiosqlite>=0.19",
    "pydantic>=2.0",
    "typer[all]>=0.9",
    "loguru>=0.7",
    "rich>=13.0",
    "pandas>=2.0",
    "pyarrow>=14.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "pytest-asyncio>=0.21", "ruff>=0.4"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E","F","I","N","UP"]
ignore = []
